model: add custom class for model and driver code.
The ideia behind CodeSnifferAI is that a LLM is able to understand the
provided code and then decide whether it contains certain code-smells.
This being said, our idea is to use the encoder of CodeT5 model in a way
that the model can be able to summarize all the info about the input
text. This is done by using the output of the model's encoder. The first
approach (the one in this commit) is to only use the hidden state of the
first token of the sequence. This is implemented in model/sniffer.py and
is tested in model/codesniffer.ipynb. We consider using all the hidden
states outputed by the encoder by using a RNN or maybe some kind of
mean.

Signed-off-by: heitorc62 <heitor.c@usp.br>